refactor: change file name convention and add tryCatch utils
