+++ src/features/login-form/index.ts
@@ -1,3 +1,3 @@
-export { default as LoginForm } from "./LoginForm";
-export { default as userReducer } from "./slice";
+export { default as LoginForm } from "./login-form";
+export { default as userReducer } from "./slice/reducer";
 export * from "./slice";
