+++ src/features/login-form/login-form.tsx
@@ -16,9 +16,10 @@ import {
 } from "@/components";
 
 import { logIn } from "./slice";
-import { useAppDispatch } from "@/hook/reduxHook";
+import { useAppDispatch } from "@/hook/redux-hook";
 import { useState } from "react";
 import { LoadingSpinner } from "@/components";
+import { tryCatch } from "@/utils";
 
 type LoginFormType = z.infer<typeof LoginSchema>;
 
@@ -37,13 +38,12 @@ const LoginForm = () => {
 
   async function onSubmit(values: LoginFormType) {
     setLoading(true);
-    try {
-      const data = await dispatch(logIn(values)).unwrap();
-      if (data.success) navigate("/");
-    } catch (error) {
-      setError(error);
-    } finally {
+
+    const { data, error } = await tryCatch(dispatch(logIn(values)).unwrap());
+    if (error) setError(error);
+    if (data?.success) {
       setLoading(false);
+      navigate("/");
     }
   }
   return (
