+++ src/hook/use-get-session.tsx
@@ -1,12 +1,12 @@
-import { getUserSession } from "@/features";
-import { useAppDispatch, useAppSelector } from "@/hook/reduxHook";
+import { getSession } from "@/features";
+import { useAppDispatch, useAppSelector } from "@/hook/redux-hook";
 import { useEffect } from "react";
 
 const useGetSession = () => {
   const dispatch = useAppDispatch();
   const data = useAppSelector((state) => state.user);
   useEffect(() => {
-    dispatch(getUserSession()).unwrap();
+    dispatch(getSession()).unwrap();
   }, []);
   return { user: data.user, loading: data.loading, error: data.error };
 };
