+++ src/pages/index.ts
@@ -1,4 +1,4 @@
-export { default as HomePage } from "./HomePage";
-export { default as LoginPage } from "./LoginPage";
+export { default as HomePage } from "./home-page";
+export { default as LoginPage } from "./login-page";
 export { default as ForgotPassword } from "./forgot-password-page";
 export { default as ResetPasswordPage } from "./reset-password-page";
