+++ src/utils/index.ts
@@ -1,1 +1,2 @@
 export { default as supabase } from "./supabase";
+export * from "./try-catch";
